Implement an alternative cancellation mechanism
Expose a thread-local
    struct musl_sig_cancel_flags {
	    sig_atomic_t volatile musl_sig_syscall_breaker;
	    sig_atomic_t          musl_sig_active;
	    uint_least64_t volatile     musl_sig_entry_mask; /*unused in libc*/
	    sigjmp_buf volatile  *musl_sig_jmpbufp; /*unused in libc*/
    };
    /*non-NULL*/ musl_sig_cancel_flags_d *musl_sig_cancel_flags(void); /*AS-safe getter*/

to the user along with

    int musl_sig_cancel__interrupt_syscall(void /*ucontext_t*/ *Uctx);

If the user handles a signal and in the handler does at least:

    void handler(int Sig, sinfo_t *Info, void *Uctx){
	musl_sig_cancel_flags()->musl_sig_syscall_breaker=Sig;
	musl_sig_cancel__interrupt_syscall(Uctx);
    }
then a delivery of that signal will pseudo-atomically break 1 next
potentially long-blocking (cancellation-point) syscall with ECANCELED,
clearing the musl_sig_sycall_breaker flag and moving it to
musl_sig_active.

This effectively ensures that a delivery of such a signal will decidedly
elicit a reaction in the target thread.

Usage:
    - clean thread cancellation without having to rewrite thread-safe single-threaded code
    in terms of pthread_cleanup_{push,pop}
    - clean thread group cancellation in response to a signal
    - partial cancellation/unwinding in reponse to a signal
    - assured reaction to signals (e.g., signals sent by other processes or timers)

TODO: architectures other than x86_64